add default photo + create any seeder

+++ app/Http/Controllers/CategoryServiceController.php
@@ -38,17 +38,17 @@ class CategoryServiceController extends Controller
 
         $request->validate([
             'name' => 'required',
-            'image' => 'required',
+            'image' => 'nullable',
         ]);
 
         if ($request->file('image')) {
             $image = $request->file('image')->store('images', 'public');
+            $categoryService->image = $image;
         }
         // dd($image);
 
         $categoryService = new CategoryService;
         $categoryService->name = $request->get('name');
-        $categoryService->image = $image;
         $categoryService->save();
 
 
@@ -84,7 +84,9 @@ class CategoryServiceController extends Controller
 
         if ($request->file('image')) {
             if ($categoryService->image && file_exists(storage_path('app/public/' . $categoryService->image))) {
+                if($categoryService->image !== 'images/serviceCategoryDefault.jpg'){
                 Storage::delete('public/' . $categoryService->image);
+                }
                 $image = $request->file('image')->store('images', 'public');
                 $categoryService->image = $image;
             }
@@ -101,7 +103,9 @@ class CategoryServiceController extends Controller
     public function destroy($idcs)
     {
         $categoryService = CategoryService::where('category_service_id', $idcs)->first();
+        if($categoryService->image !== 'images/serviceCategoryDefault.jpg'){
         Storage::delete('public/' . $categoryService->image);
+        }
         $categoryService->delete();
         return redirect()->route('categoryService.index')
             ->with('success', 'Category Service seccesfully Deleted');
