+++ app/Http/Controllers/UserController.php
@@ -11,11 +11,7 @@ use Auth;
 
 class UserController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    
     public function index(Request $request)
     {
         $search = $request->get('search');
@@ -27,11 +23,6 @@ class UserController extends Controller
         return view('admin.adminIndex', compact('admins'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
 
 
     public function profile($iduser)
@@ -80,12 +71,7 @@ class UserController extends Controller
         //
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+    
     public function store(Request $request)
     {
         $request->validate([
@@ -109,35 +95,19 @@ class UserController extends Controller
             ->with('success', 'New Admin Added Succesfully');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
+    
     public function show(User $user)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
+    
     public function edit(User $admin)
     {
         return view('admin.adminEdit', ['admin' => $admin]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
+   
     public function update(Request $request, User $admin)
     {
         $request->validate([
@@ -148,10 +118,12 @@ class UserController extends Controller
         ]);
         if ($request->file('image')) {
             if ($admin->image) {
-                Storage::delete('public/' . $admin->image);
+                if($admin->image !== 'images/userDefault.jpg'){
+                    Storage::delete('public/' . $admin->image);
+                }
+                $image = $request->file('image')->store('images', 'public');
+                $admin->image = $image;
             }
-            $image = $request->file('image')->store('images', 'public');
-            $admin->image = $image;
         }
 
         $admin->name = $request->get('name');
@@ -162,15 +134,12 @@ class UserController extends Controller
             ->with('success', 'Admin seccesfully Updated');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
+    
     public function destroy(User $admin)
     {
+        if($admin->image !== 'images/userDefault.jpg'){
         Storage::delete('public/' . $admin->image);
+        }
         $admin->delete();
         return redirect()->route('admins.index')
             ->with('success', 'Admin seccesfully Deleted');
